// apps/web/src/api/models/summaryResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PayPay Money Diary API
 * PayPay 家計簿アプリケーション API
 * OpenAPI spec version: 1.0.0
 */
import type { SummaryResponseSummary } from "./summaryResponseSummary";
import type { CategoryBreakdown } from "./categoryBreakdown";
import type { MonthlyBreakdown } from "./monthlyBreakdown";

export interface SummaryResponse {
  summary: SummaryResponseSummary;
  /** カテゴリ別内訳 */
  categoryBreakdown: CategoryBreakdown[];
  /** 月別内訳 */
  monthlyBreakdown: MonthlyBreakdown[];
}