// apps/web/src/api/models/categoryBreakdown.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PayPay Money Diary API
 * PayPay 家計簿アプリケーション API
 * OpenAPI spec version: 1.0.0
 */

export interface CategoryBreakdown {
  /**
   * カテゴリID
   * @nullable
   */
  categoryId: string | null;
  /** カテゴリ名 */
  categoryName: string;
  /** カテゴリ色 */
  categoryColor: string;
  /**
   * カテゴリアイコン
   * @nullable
   */
  categoryIcon: string | null;
  /** 合計金額 */
  totalAmount: number;
  /** 取引数 */
  transactionCount: number;
}